correspondent_network/ingestion_service/requirements.txt
pytesseract==0.3.10
Pillow==10.2.0
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
